requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==23.1.0
aiogram==2.23.1
aiohttp==3.8.4
aiormq==6.4.2
aiosignal==1.3.1
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bidict==0.22.1
boto3==1.26.74
botocore==1.29.74
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==1.9.2
cryptography==39.0.1
cycler==0.11.0
dask==2022.10.2
dnspython==1.16.0
docopt==0.6.2
exceptiongroup==1.1.1
faiss-cpu==1.7.3
fbmessenger==6.0.0
filelock==3.11.0
fire==0.5.0
firebase-admin==6.1.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.79.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpcio==1.51.3
grpcio-status==1.51.3
h5py==3.8.0
httplib2==0.21.0
httptools==0.5.0
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
jax==0.4.8
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==4.16.0
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
locket==1.0.0
magic-filter==1.0.9
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
mattermostwrapper==2.2
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.4
multidict==5.2.0
networkx==2.6.3
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.3.0
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.28
proto-plus==1.22.2
protobuf==4.22.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
pytest==7.3.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.5.4
rasa-sdk==3.5.1
redis==4.5.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
sentry-sdk==1.11.1
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.20.0
SQLAlchemy==1.4.46
sympy==1.11.1
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
transformers==4.28.0
twilio==7.14.2
typeguard==2.13.3
typing-utils==0.1.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
